requirements.txt
requests
python-dotenv
pandas
ruff
pytest
pre-commit
